hook for updateproblemschema
