+++ tworaven_apps/ta2_interfaces/templates/test_responses/updateproblemschema_ok.json
@@ -1,0 +1,6 @@
+{
+  "status": {
+    "code": "OK",
+    "details": "Problem Schema Updated"
+  }
+}
